requirements.txt
Django==5.1.4
gunicorn==23.0.0
psycopg2-binary==2.9.10
dj-database-url==2.3.0
python-dotenv==1.0.1
whitenoise==6.8.2
Pillow==11.0.0
python-dateutil==2.9.0
